requirements.txt
numpy>=1.26.0
pandas>=2.2.0
pyarrow>=16.0.0
python-dotenv>=1.0.0
pytz>=2024.0
requests>=2.30.0
rich>=13.0.0
scikit-learn>=1.6.0
setuptools>=69.0.0
tabulate>=0.9.0
urllib3>=2.0.0
yfinance>=0.2.40
PyYAML>=6.0.0
pandas-datareader
colorama
matplotlib>=3.10.0
seaborn>=0.13.0
pytest>=7.0.0
pdfplumber>=0.11.0
pytest-flask>=1.3.0
faker>=24.0.0
sqlalchemy>=2.0.0